requirements.txt
Django
django-bootstrap4
django-taggit
psycopg2
